flask/memo-app-db/requirement.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
